requirements.txt
# ── Core ML ───────────────────────────────────────────────────────────────────
torch>=2.2.0
torchaudio>=2.2.0
torchvision>=0.17.0
numpy>=1.24.0
scipy>=1.10.0

# ── Audio I/O & DSP ──────────────────────────────────────────────────────────
librosa>=0.10.0
soundfile>=0.12.1
audioread>=3.0.0
praat-parselmouth>=0.4.3        # F0 extraction via Praat
pydub>=0.25.1

# ── Speech Models ─────────────────────────────────────────────────────────────
openai-whisper>=20231117        # Whisper v3
transformers>=4.38.0            # HuggingFace (wav2vec2, SpeechBrain wrappers)
speechbrain>=1.0.0              # x-vector / d-vector embeddings
TTS>=0.22.0                     # Coqui TTS (VITS / YourTTS)

# ── Denoising ─────────────────────────────────────────────────────────────────
deepfilternet>=0.5.6            # DeepFilterNet denoiser (fallback: spectral sub)

# ── Phonetics / IPA ───────────────────────────────────────────────────────────
phonemizer>=3.2.1               # espeak-ng backend
epitran>=1.22                   # G2P for Hindi (hin-Deva)
gruut>=2.2.3                    # G2P fallback

# ── NLP / N-gram LM ──────────────────────────────────────────────────────────
nltk>=3.8.1
kenlm                           # KenLM Python bindings (pip install https://...)
# If kenlm fails: pip install https://github.com/kpu/kenlm/archive/master.zip

# ── Data & Utilities ──────────────────────────────────────────────────────────
yt-dlp>=2024.1.0                # YouTube download
tqdm>=4.66.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.13.0
scikit-learn>=1.3.0
jiwer>=3.0.3                    # WER computation
pymcd                           # Mel-Cepstral Distortion

# ── Adversarial ───────────────────────────────────────────────────────────────
# (FGSM implemented from scratch with PyTorch autograd — no extra library needed)
